Lots! Mainly getting data to TimerActivity.kt though. Android studios crashed and took my nice verbose comment with it so this it what you get.

+++ app/src/main/java/com/kth/id2216/group3/time/ui/createtimer/CreateTimerViewModel.kt
@@ -22,7 +22,7 @@ class CreateTimerViewModel @Inject constructor(
     val timers: LiveData<List<Timer>> = timerRepository.getAll().asLiveData()
     val categories: LiveData<List<Category>> = categoryRepository.getAll().asLiveData()
 
-    fun addTimer(name: String, goal: Int) {
+    fun addTimer(name: String, goal: Double) {
         viewModelScope.launch {
             val timer = Timer(name = name, goal = goal)
             timerRepository.insert(timer)
